requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
requests==2.21.0
urllib3==1.24.1
Werkzeug==0.14.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dialogflow==0.5.2
eventlet==0.24.1
Flask==1.0.2
flask-socketio==3.3.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
redis==3.2.1
requests==2.21.0
urllib3==1.24.1
Werkzeug==0.14.1
